Moves search stats into mailbox, fixes #244
